requirements.txt
# Core dependencies
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
urllib3==2.0.4
dnspython==2.4.2
cryptography==41.0.4
Pillow==10.0.1
python-whois==0.8.0
validators==0.22.0

# Web scraping and automation
selenium==4.15.0
webdriver-manager==4.0.1

# Data analysis and ML
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2

# Image processing
opencv-python==4.8.1.78
imagehash==4.3.1

# Security and SSL
pyOpenSSL==23.2.0
certifi==2023.7.22

# API and database
# sqlite3 is built into Python
sqlalchemy==2.0.21
marshmallow==3.20.1

# CLI and utilities
click==8.1.7
colorama==0.4.6
tqdm==4.66.1
python-dotenv==1.0.0

# Testing and development
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0

# Optional ML dependencies
# tensorflow==2.13.0
# torch==2.0.1
